Modify architecture + fix import cycle
