+++ cmd/utask/root.go
@@ -20,9 +20,9 @@ import (
 	"github.com/ovh/utask/api"
 	"github.com/ovh/utask/db"
 	"github.com/ovh/utask/engine"
-	"github.com/ovh/utask/models/hook"
 	"github.com/ovh/utask/models/tasktemplate"
 	"github.com/ovh/utask/pkg/auth"
+	"github.com/ovh/utask/pkg/hook"
 	notify "github.com/ovh/utask/pkg/notify/init"
 	"github.com/ovh/utask/pkg/plugins"
 	"github.com/ovh/utask/pkg/plugins/builtin"
